+++ src/tests/monster/aura.c
@@ -22,6 +22,8 @@
 int setup_tests(void **state) {
 	set_file_paths();
 	init_angband();
+
+	(void) test_seed_rng_reported(suite_name);
 	*state = NULL;
 	return 0;
 }