+++ src/tests/monster/restele.c
@@ -21,6 +21,8 @@
 int setup_tests(void **state) {
 	set_file_paths();
 	init_angband();
+
+	(void) test_seed_rng_reported(suite_name);
 	*state = NULL;
 	return 0;
 }